pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastmcp>=2.10.5",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "pymilvus>=2.5.12",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
